requirements.txt
fastapi
uvicorn
mysql_connector
requests
python-dotenv